printf %f

+++ src/inttypes.h
@@ -14,8 +14,8 @@ typedef unsigned char       uint8_t;
 typedef unsigned short      uint16_t;
 typedef unsigned int        uint32_t;
 typedef unsigned long long  uint64_t;
-typedef int32_t             ssize_t;
-typedef uint32_t            size_t;
+typedef long                ssize_t;
+typedef unsigned long       size_t;
 typedef int64_t             intmax_t;
 typedef uint64_t            uintmax_t;
